requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
crc32c==2.2.post0
cycler==0.10.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.21.0
oauthlib==3.1.1
Pillow==8.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
rawutil==2.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
threadpoolctl==2.1.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
